+++ lib/crateNameFromCargoToml.nix
@@ -13,7 +13,7 @@ let
   '';
 
   src = args.src or throwMsg;
-  cargoToml = args.cargoToml or (args.src + "/Cargo.toml");
+  cargoToml = args.cargoToml or (src + "/Cargo.toml");
   cargoTomlContents = args.cargoTomlContents or (
     if builtins.pathExists cargoToml
     then builtins.readFile cargoToml
@@ -33,19 +33,24 @@ let
     `NIX_ABORT_ON_WARN=1 nix --option pure-eval false --show-trace ...`
   '';
 
-  traceMsg = tomlName: drvName: placeholder: lib.flip lib.trivial.warn placeholder ''
+  traceMsg = tomlName: drvName: placeholder: workspaceHints: lib.flip lib.trivial.warn placeholder ''
     crane will use a placeholder value since `${tomlName}` cannot be found in ${debugPath}
     to silence this warning consider one of the following:
     - setting `${drvName} = "...";` in the derivation arguments explicitly
-    - setting `package.${tomlName} = "..."` or `workspace.package.${tomlName} = "..."` in the root Cargo.toml
+    - setting `package.${tomlName} = "..."` or ${lib.concatStringsSep " or " workspaceHints} in the root Cargo.toml
     - explicitly looking up the values from a different Cargo.toml via 
       `craneLib.crateNameFromCargoToml { cargoToml = ./path/to/Cargo.toml; }`
     ${hint}
   '';
 
-  internalName = internalCrateNameFromCargoToml toml;
+  internalName = internalCrateNameFromCargoToml toml debugPath;
 in
 {
-  pname = internalName.pname or (traceMsg "name" "pname" "cargo-package");
-  version = internalName.version or (traceMsg "version" "version" "0.0.1");
+  pname = internalName.pname or (traceMsg "name" "pname" "cargo-package" [
+    ''`package.metadata.crane.name` = "..."''
+    ''`workspace.metadata.crane.name` = "..."''
+  ]);
+  version = internalName.version or (traceMsg "version" "version" "0.0.1" [
+    ''`workspace.package.version` = "..."''
+  ]);
 }